+++ src/egtgt2.c
@@ -158,7 +158,7 @@ int16 computeTargetBearing(int16 targetX, int16 targetY, int16 wantBearing) {
 
 // ==== seg000:0xc82d ====
 int16 computeLoftAngle() {
-    return (int16)((unsigned long)((long)(0x4000 - abs(g_ourPitch)) << 12) / (unsigned long)(uint16)(g_viewZ + 0x1000)) - 0x4000;
+    return (int16)((uint32)((long)(0x4000 - abs(g_ourPitch)) << 12) / (uint32)(uint16)(g_viewZ + 0x1000)) - 0x4000;
 }
 
 // ==== seg000:0xc864 ====
